macaroon_recipes: add MacaroonRecipe function
To make it easy to bake custom macaroons for just the subservers that
lndclient supports, we add a function that returns all required macaroon
permissions for a given RPC package.
This function uses the ListPermissions RPC of lnd to query all existing
RPC method URIs and their required permissions. With some reflection
magic a subset of permissions is them assembled for the given package.