client: add option to block until lnd is unlocked
This commit adds an optional wait for lnd to unlock to LndServices.
The wallet unlocker is not used, because polling the unlock endpoint
exacerbates a known deadlock:
https://github.com/lightningnetwork/lnd/issues/3631.

Instead, a call to the main grpc server is used. The wallet is
considered locked when we receive a grpc unimplemented error, because
the main server does not become active until the wallet is unlocked.
Once the wallet is unlocked, there is a race condition where a query
to the main server can return an unavailable code while the server is
busy registering. We use the grpc wait until ready error which will
allow the call to wait until our wait interval has elapsed before it
fails. This allows lnd some time to come up.